backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.25
alembic==1.13.1

# LangChain & LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-deepseek>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-google-genai>=0.3.0
langgraph>=0.2.0
